requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
aim==3.3.5
coloredlogs==15.0.1
efficientnet_pytorch==0.7.1
h5py==3.1.0
humanize==4.2.3
importlib_metadata==3.10.0
ipympl
ipython==7.16.1
jupyterlab==3.1.7
libauc==1.1.5
matplotlib==3.4.2
nltk==3.6.7
numpy==1.19.2
opencv_python==4.5.3.56
openpyxl==3.0.10
pandas==1.3.2
Pillow==9.0.0
plotext==3.1.3
pre-commit==2.20.0
pytest==7.1.2
pytest-cov==3.0.0
pytorch-lightning==1.6.5
PyYAML==6.0
recommonmark==0.6.0
scikit_learn==1.0.2
scipy==1.6.2
seaborn==0.11.2
setuptools==59.6.0
spacy==3.2.1
Sphinx==4.0.2
tensorboard==2.6.0
timm==0.4.12
torch==1.10.2+cu113
torchvision==0.11.3+cu113
tox==3.25.1
tqdm==4.62.1
transformers==4.12.5
wandb==0.12.21
